p2v: Send physical server 'dmesg' output to debug dir on conversion server.
If the virt-p2v ISO doesn't contain the correct drivers for the
hardware on the physical server, that can cause peculiar conversion
problems.  This was hard to diagnose because we could not see the
'dmesg' output from the physical server.

Upload 'dmesg' output from the physical server (more precisely: from
the virt-p2v ISO's kernel running on the physical server) to the
virt-v2v conversion server debugging directory, so that this
information is available for debugging later.

Unfortunately this involves yet another copy of the 'read_whole_file'
function.