ja: Fix unclosed L<> which breaks POD.
Since commit b23e1497747f1cdfdedfd0567e7748fdc3ae0684, we have been
strict about POD errors.  An error in the Japanese translation caused
"L</guestfish" (ie. an unclosed L<>) to appear in the POD.

This commit fixes the problem, but when building you may need to do:

  rm po-docs/ja/guestfish.pod
  rm po-docs/podfiles; make -C po-docs update-po

in order to update the broken po-docs/ja/guestfish.pod file.

Unfortunately I have been unable to upload this change back to the
Zanata servers because of a variety of client problems.