daemon: Choose /usr/sbin first for the daemon (debian bug 838995).
Debian is also doing a UsrMove/UsrMerge:
https://wiki.debian.org/UsrMerge

However it is not finalized that Debian will actually make this
change.

Since some Debian systems have /sbin as a symlink and other have /sbin
as a real directory, and we should avoid choosing a symlinked
directory for the daemon, the easiest fix is simply to probe /usr/sbin
before /sbin since under all scenarios (and Fedora too) /usr/sbin is a
real directory.

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=838995