po-docs: Fix mistranslated message in Ukrainian file.
A message was translated without the trailing \n.  As it was in a
verbatim section of POD, I actually reverted the whole translation.

The error was:

  ./uk.po:53533: 'msgid' and 'msgstr' entries do not both end with '\n'
  msgfmt: found 1 fatal error

In fact po4a (or whatever calls msgfmt) seems to ignore the error and
continues which is another problem.

Fixes commit 794b6a1acfb3af96caf901d64c9658052b3d5438.