build: build mlaugeas with -Wno-shift-negative-value
The embedded copy of ocaml-augeas does Val_int(-1), which in turns
triggers warnings in newer GCC versions about "left shift of negative
value".  The issue actually lies in the OCaml headers (mlvalues.h in
particular), and it was fixed in newer OCaml versions.

Since the code is actually correct, disable -Wshift-negative-value with
-Wno-shift-negative-value (checking whether the compiler has it).