po-docs/podfiles: Don't add tarball files to po-docs
When we are building the tarball (ie. make dist) we may also update
podfiles.  However while this is happening there can be a directory
called guestfs-tools-1.XXX in the current directory, and we were
inadvertantly adding files from that directory to podfiles.

There was already a regexp meant to avoid this, fix it.

Fixes: commit 4354a3126152a2748cc9097cba139b3908ccc342