win-reg: Fix output of virt-win-reg --version
Before we split libguestfs and guestfs-tools the version of programs
was exactly the same as the version of libguestfs.  This was
convenient for virt-win-reg where it is difficult to encode the
version into the script, because virt-win-reg could simply use the
version from Sys::Guestfs (ie. libguestfs).  However after the split
this no longer applies, so we must do the difficult thing.

After this change virt-win-reg will print both the version of
guestfs-tools and the version of libguestfs, eg:

    $ virt-win-reg --version
    virt-win-reg 1.47.1
    libguestfs 1.45.5fedora=35,release=1.fc35,libvirt

Reported-by: Yongkui Guo
Fixes: commit 4354a3126152a2748cc9097cba139b3908ccc342
Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1961160
(cherry picked from commit 9e9eeeaf4c4b478eca82bf0d4c3e680a624d6d91)