templates: Add CentOS Stream 8
This adds CentOS Stream 8 to the set of images that the builder
can use. CentOS Stream 8 is a continuously delivered upstream of
the next RHEL 8.x release. It doesn't have a minor version number.

It appears that my index-fragment has a hash in it, and thus this
commit may need to be amended by the maintainer before merging
into the main repo.

This has been tested by creating a local centosstream-8.xz image
then consuming that image via a local repos.d file and using
virt-builder to build a bootable CentOS Stream 8 image:

     $ cat ~/.config/virt-builder/repos.d/local.conf
     [local]
     uri=file:///data/src/guestfs-tools/builder/templates/index
     $ virt-builder --list | grep stream
     centosstream-8           x86_64     CentOS Stream 8
     $ virt-builder centosstream-8
     [   1.0] Downloading: file:///data/src/guestfs-tools/build...
     [   1.7] Planning how to build this image
     [   1.7] Uncompressing
     [   7.8] Opening the new disk
     [  12.6] Setting a random seed
     [  12.6] Setting passwords
     virt-builder: Setting random password of root to qmjtBFyLX...
     [  13.5] Finishing off
                   Output file: centosstream-8.img
                   Output size: 6.0G
                 Output format: raw
            Total usable space: 5.4G
                    Free space: 3.8G (71%)