customize: Add support for the yescrypt hash method.
The yescrypt hashing method is considered to be much stronger
than sha512crypt and fully supported by libxcrypt >= 4.3.

It is based on NIST-approved primitives, like PBKDF2, and on
par with argon2 in strength.

Fresh installed systems, as well as newly computed hashes for
the UNIX shadow file should prefer this method, if the crypt()
function, the distro ships, is capable of computing such hashes.

Signed-off-by: Björn Esser <besser82@fedoraproject.org>