sparsify: Deconflict signal handlers from [On_exit] with this tool
We cannot be sure that the [On_exit] module has or has not registered
a signal handler.  If not, then the signal handlers registered by
virt-sparsify would be overwritten (versus overwriting the ones
registered by [On_exit] which is what we want).  So expose the
[On_exit.register] function and call it.

Luckily virt-sparsify is the only tool which tries to register a
signal handler, so this complexity is confined in a single place.