Ensure the machine-id operation is the last one
We must ensure the modified image doesn't have any content in
/etc/machine-id file. This will then trigger the usual
systemd-machine-id-setup during the first boot of the image, ensuring
we'll get an as unique as possible ID.

This is mandatory, since the tools are generating a machine-id at
different stages, for instance in order to ensure packages are
installing without any issue[1]

Of course, if the operator wants to get an actual machine-id, they can
just disable the "machine-id" operation. Note that it's enabled by
default, so the issue is more about ordering than anything.

[1] For instance here: https://github.com/rwmjones/guestfs-tools/blob/3a498512f58bc431db490e96cdb712b19389bda4/customize/customize_run.ml#L170-L188