test-data/phony-guests: Allow virt-v2v to work against phony Fedora
We didn't use the phony Fedora guest before with virt-v2v (only the
phony Windows image).  This commit makes miscellaneous changes so that
it can be used for testing:

 - Add dummy rpm and dracut commands.

 - Add dummy kernel, initramfs and modules directory.

 - Add dummy grub configuration pointing to the kernel.

RWMJ:

I had to make a change to the virt-ls test since it depends on the
exact contents of /bin and /boot in the Fedora test image.

Cherry-picked-from: virt-v2v commit 1e75569aa074a50e96867d1021651ca2f75bcc16