make-fs: Use -S option with -z
Basically the file -z option is broken for distributions that enable
seccomp support in the 'file' command.  Turn off seccomp when we know
we'll be using -z.

We hit this before in the libguestfs daemon:
https://github.com/libguestfs/libguestfs/issues/100

It was reported upstream but the fix was rejected:
https://bugs.astron.com/view.php?id=406

See also libguestfs commit 23986d3c4f4d1f9cbac44cc743d3e6af721e4237
("file: Use -S option with -z").

Note this requires file > 5.32.  We could make it work for earlier
versions of file with some extra work:
https://listman.redhat.com/archives/libguestfs/2023-September/032613.html
& libguestfs commit c95d8c4cf64142bb707b42c32cf3e1ba3c4a5eb1

Reported-by: Jax Young
Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2266201