sparsify: Remove test-virt-sparsify-in-place-fstrim-unsupported.sh
With kernel 6.11.0 this test suddenly started to work, as if the minix
filesystem in this kernel suddenly supports trimming.  I checked the
kernel source and I can't find anything relevant so I'm not sure what
happened.  In any case this test is of very marginal utility so just
delete it.