builder: Link to the actually existing virt-builder templates dir
This local virt-builder conf file is used when you run the tests (with
the exception noted below) or use the ./run script directly.  The
./run script sets VIRT_BUILDER_DIRS to point here.

The previous contents of this file pointed to a directory which did
not exist -- it was part of libguestfs and didn't get copied into the
new guestfs-tools source tree.  This now points to the
builder/templates/ subdirectory.  This is enough to get certain
commands like './run virt-builder --list' to work, enough to run the
virt-builder tests, but it does NOT contain any installable templates.

Note 'make check-slow' in builder/ unsets $VIRT_BUILDER_DIRS, and thus
uses the public virt-builder repositories (requiring network access,
but giving access to the full range of real, installable templates).