test-data/phony-guests: Increase size of Windows image
For testing I tried to upload a very large Windows registry to the
phony Windows image.  However because of the 256M limit on the size of
the second partition (and NTFS overhead?) this was not possible.

There's no real penalty to increasing the size of this image since
it's just virtual space, so increase the virtual size to 2G (256M/1.7G
split).